+++ tools/cgquery/include/Utils.h
